fix: product price, stock type 변경 #150
- Error 코드 추가
- Long -> Integer

+++ product-service/src/main/java/com/oingmaryho/business/productservice/application/dto/request/ProductCreateRequestServiceDto.java
@@ -8,7 +8,7 @@ public record ProductCreateRequestServiceDto(
 	String productCode,
 	String name,
 	UUID manageHubId,
-	Long stock,
-	Long price
+	Integer stock,
+	Integer price
 ) {
 }
