+++ product-service/src/main/java/com/oingmaryho/business/productservice/application/dto/request/ProductUpdateRequestServiceDto.java
@@ -6,7 +6,7 @@ public record ProductUpdateRequestServiceDto(
 	UUID id,
 	String companyName,
 	String name,
-	Long price,
-	Long stock
+	Integer price,
+	Integer stock
 ) {
 }
