+++ product-service/src/main/java/com/oingmaryho/business/productservice/presentation/dto/request/ProductUpdateRequestDto.java
@@ -3,8 +3,8 @@ package com.oingmaryho.business.productservice.presentation.dto.request;
 public record ProductUpdateRequestDto(
 	String companyName,
 	String name,
-	Long price,
-	Long stock
+	Integer price,
+	Integer stock
 
 ) {
 }
